requirements.server.txt
streamlit>=1.29.0
openai>=1.3.0
httpx>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
